Fix download files endpoint when impersonating user
